+++ .env.example
@@ -1,0 +1,2 @@
+KEY_MNEMONIC=""
+KEY_INFURA_API_KEY=""
