+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules/
 build/
 contracts/.flattened
+.openzeppelin/.session
+.env
