+++ .waffle.json
@@ -17,7 +17,7 @@
     "evmVersion": "istanbul",
     "optimizer": {
       "enabled": true,
-      "runs": 999999
+      "runs": 200
     }
   }
 }
