+++ test/DXswapPair.spec.ts
@@ -178,7 +178,7 @@ describe('DXswapPair', () => {
     await mineBlock(provider, (await provider.getBlock('latest')).timestamp + 1)
     const tx = await pair.swap(expectedOutputAmount, 0, wallet.address, '0x', overrides)
     const receipt = await tx.wait()
-    expect(receipt.gasUsed).to.eq(75927)
+    expect(receipt.gasUsed).to.eq(76461)
   })
 
   it('burn', async () => {
