+++ truffle-config.js
@@ -1,3 +1,9 @@
+const HDWalletProvider = require('truffle-hdwallet-provider')	
+require('dotenv').config();
+
+mnemonic = process.env.KEY_MNEMONIC;
+infuraApiKey = process.env.KEY_INFURA_API_KEY;
+
 module.exports = {	
   networks: {	
     rpc: {	
@@ -13,7 +19,39 @@ module.exports = {
       port: 8545,	
       gas: 9000000,	
       gasPrice: 10000000000 //10 Gwei	
-    },
+    },	
+    mainnet: {	
+      provider: function () {	
+        return new HDWalletProvider(mnemonic, `https://mainnet.infura.io/v3/${infuraApiKey}`)	
+      },	
+      network_id: '1',	
+      gas: 9000000,	
+      gasPrice: 10000000000 //10 Gwei	
+    },	
+    rinkeby: {	
+      provider: function () {	
+        return new HDWalletProvider(mnemonic, `https://rinkeby.infura.io/v3/${infuraApiKey}`)	
+      },	
+      network_id: '4',	
+      gas: 9000000,	
+      gasPrice: 10000000000 //10 Gwei	
+    },	
+    ropsten: {	
+      provider: function () {	
+        return new HDWalletProvider(mnemonic, `https://ropsten.infura.io/v3/${infuraApiKey}`)	
+      },	
+      network_id: '3',	
+      gas: 8000000,	
+      gasPrice: 10000000000 //10 Gwei	
+    },	
+    kovan: {	
+      provider: function () {	
+        return new HDWalletProvider(mnemonic, `https://kovan.infura.io/v3/${infuraApiKey}`)	
+      },	
+      network_id: '42',	
+      gas: 9000000,	
+      gasPrice: 10000000000 //10 Gwei	
+    }	
   },	
   build: {},	
   compilers: {
